--- src/main/java/hr/bitnaprednost/days/D02_RockPaperScissors/model/Outcome.java
@@ -1,7 +1,0 @@
-package hr.bitnaprednost.days.D02_RockPaperScissors.model;
-
-public enum Outcome {
-    DRAW,
-    WIN,
-    LOSS
-}
