--- src/main/java/hr/bitnaprednost/days/D02_RockPaperScissors/model/Shape.java
@@ -1,7 +1,0 @@
-package hr.bitnaprednost.days.D02_RockPaperScissors.model;
-
-public enum Shape {
-    ROCK,
-    PAPER,
-    SCISSOR
-}
