+++ src/main/java/hr/bp/aoc/days/rockPaperScissors/model/Outcome.java
@@ -1,0 +1,7 @@
+package hr.bp.aoc.days.rockPaperScissors.model;
+
+public enum Outcome {
+    DRAW,
+    WIN,
+    LOSS
+}
