+++ src/main/java/hr/bp/aoc/days/rockPaperScissors/model/Shape.java
@@ -1,0 +1,7 @@
+package hr.bp.aoc.days.rockPaperScissors.model;
+
+public enum Shape {
+    ROCK,
+    PAPER,
+    SCISSOR
+}
